# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Soiree API
  version: 0.1.0
  description: |
    Soiree — Robocon test-run video manager HTTP API.

    Phase 1 の主要リソース（masters / sessions / videos / runs / markers / search）を契約として定義する。
    実装は spec-devflow.md §4〜§9 の順に handler / service / UI を縦断で積む。

    ## 共通方針

    - JSON フィールド名は camelCase
    - ID は UUID 文字列
    - 時刻は ISO 8601 (RFC 3339) タイムスタンプ
    - 動画上の時間は秒単位（int = 整数秒、float = 小数秒）
    - リスト系は `{ data: [...], pagination: {...} }` のエンベロープ
    - ページングは cursor ベース (`?cursor=...&limit=...`)
    - エラーは `ErrorResponse` で統一
servers:
  - url: http://localhost:8080
    description: Local development
  - url: /
    description: Same-origin (production / docker)
security:
  - xUserId: []
...
paths:
...
  /devices:
    get:
      tags: [devices]
      operationId: listDevices
      summary: 機材一覧
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceList"
    post:
      tags: [devices]
      operationId: createDevice
      summary: 機材作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
  /devices/{deviceId}:
    parameters:
      - $ref: "#/components/parameters/DeviceId"
    get:
      tags: [devices]
      operationId: getDevice
      summary: 機材取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [devices]
      operationId: updateDevice
      summary: 機材更新（時計オフセット含む）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [devices]
      operationId: deleteDevice
      summary: 機材削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Teams ----------
...
  /sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      operationId: getSession
      summary: セッション取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [sessions]
      operationId: updateSession
      summary: セッション更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [sessions]
      operationId: deleteSession
      summary: セッション削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /videos:
    get:
      tags: [videos]
      operationId: listVideos
      summary: 動画一覧
      parameters:
        - name: tournamentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: deviceId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: unassigned
          in: query
          description: true なら sessionId が未設定の動画のみ
          required: false
          schema:
            type: boolean
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoList"
  /videos/{videoId}:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [videos]
      operationId: getVideo
      summary: 動画取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [videos]
      operationId: updateVideo
      summary: 動画メタデータ更新（recorded_at / time_offset_sec / session 紐付け）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [videos]
      operationId: deleteVideo
      summary: 動画削除（オブジェクトストレージのファイルも削除）
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /runs/{runId}/videos:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      tags: [runs]
      operationId: addRunVideo
      summary: Run にアングル動画を紐付ける
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRunVideoRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunVideo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /runs/{runId}/videos/{runVideoId}:
    parameters:
      - $ref: "#/components/parameters/RunId"
      - $ref: "#/components/parameters/RunVideoId"
    patch:
      tags: [runs]
      operationId: updateRunVideo
      summary: アングル動画の offset / angle ラベル更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRunVideoRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunVideo"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [runs]
      operationId: removeRunVideo
      summary: Run からアングル動画を外す
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    xUserId:
      type: apiKey
      in: header
      name: X-User-Id
      description: |
        現在ユーザーの UUID。セルフホスト・全員同権限の前提のため
        Phase 1 はクライアントが選択ユーザーを header で送る簡易方式。
        将来正式な認証を入れたら securityScheme を差し替える。
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: 前回レスポンスの `pagination.nextCursor`
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: 1 レスポンスあたり最大件数（1〜200、既定 50）
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VideoId:
      name: videoId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunVideoId:
      name: runVideoId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: 競合（重複や状態不一致）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 機械可読なエラーコード
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Pagination:
      type: object
      required: [hasMore]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: [string, "null"]
          description: 次ページを取得するための cursor

    # ---------- Enums ----------
    SessionModeHint:
      type: string
      enum: [practice, pre_match]
      description: Session の表示モードヒント
...
    Device:
      type: object
      required: [id, name, defaultTimeOffsetSec, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        defaultTimeOffsetSec:
          type: number
          description: 機材の時計ズレ（秒、小数可）。プラスなら機材時刻が進んでいる
        createdAt:
          type: string
          format: date-time
    CreateDeviceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        defaultTimeOffsetSec:
          type: number
          default: 0
    UpdateDeviceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        defaultTimeOffsetSec:
          type: number
    DeviceList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ---------- Team ----------
...
    Session:
      type: object
      required: [id, name, modeHint, tournamentId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        startedAt:
          type: [string, "null"]
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
        location:
          type: [string, "null"]
        modeHint:
          $ref: "#/components/schemas/SessionModeHint"
        tournamentId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    UpdateSessionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        startedAt:
          type: [string, "null"]
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
        location:
          type: [string, "null"]
        modeHint:
          $ref: "#/components/schemas/SessionModeHint"
        tournamentId:
          type: string
          format: uuid
...
    Video:
      type: object
      required: [id, tournamentId, storageKey, displayName, timeOffsetSec, effectiveRecordedAt, hasThumbnail, hlsStatus, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
        sessionId:
          type: [string, "null"]
          format: uuid
        deviceId:
          type: [string, "null"]
          format: uuid
        uploaderId:
          type: [string, "null"]
          format: uuid
        storageKey:
          type: string
          description: オブジェクトストレージのキー
        displayName:
          type: string
          description: UI 表示用の名前。初期値はアップロード時のファイル名
        recordedAt:
          type: [string, "null"]
          format: date-time
          description: 撮影時刻の生値（ffprobe 抽出または手入力）。時刻補正は未適用
        effectiveRecordedAt:
          type: [string, "null"]
          format: date-time
          description: |
            時刻補正を適用した実時刻 = recordedAt - (Device.defaultTimeOffsetSec
            + timeOffsetSec)。タイムライン配置・重複判定・Session マッチングに
            使う。recordedAt が null のときは null。
        durationSec:
          type: [integer, "null"]
        timeOffsetSec:
          type: number
          description: 個別オフセット（Device default に対する追加補正、秒、小数可）
        hasThumbnail:
          type: boolean
          description: サムネイルが生成済みなら true。 GET /videos/{id}/thumbnail-url で取得
        hlsStatus:
          type: string
          description: |
            HLS 変換パイプラインの状態。`ready` のときのみ playback-url が
            HLS (kind=hls) を返す。それ以外は MP4 フォールバック。
          enum: [pending, planning, encoding, ready, failed]
        createdAt:
          type: string
          format: date-time
    UpdateVideoRequest:
      type: object
      properties:
        sessionId:
          type: [string, "null"]
          format: uuid
        deviceId:
          type: [string, "null"]
          format: uuid
        recordedAt:
          type: [string, "null"]
          format: date-time
        timeOffsetSec:
          type: number
        displayName:
          type: string
    VideoList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    RunVideo:
      type: object
      required: [id, runId, videoId, videoOffsetStartSec, videoOffsetEndSec, runOffsetSec, angleLabel]
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        videoId:
          type: string
          format: uuid
        videoOffsetStartSec:
          type: integer
          description: Source video 側で「Run の中で見せる」区間の開始秒
        videoOffsetEndSec:
          type: integer
          description: Source video 側で「Run の中で見せる」区間の終了秒
        runOffsetSec:
          type: integer
          minimum: 0
          description: Run タイムライン上でこの動画が登場する秒。0 なら Run 開始と同時
        angleLabel:
          type: string
          description: 例 "正面" / "コート横" / "drone"
    AddRunVideoRequest:
      type: object
      required: [videoId, videoOffsetStartSec, videoOffsetEndSec]
      properties:
        videoId:
          type: string
          format: uuid
        videoOffsetStartSec:
          type: integer
        videoOffsetEndSec:
          type: integer
        runOffsetSec:
          type: integer
          minimum: 0
          default: 0
        angleLabel:
          type: string
          default: ""
    UpdateRunVideoRequest:
      type: object
      properties:
        videoOffsetStartSec:
          type: integer
        videoOffsetEndSec:
          type: integer
        runOffsetSec:
          type: integer
          minimum: 0
        angleLabel:
          type: string

    # ---------- Marker ----------
...